requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
Pillow==10.3.0
tensorflow==2.16.1
numpy==1.26.4
requests==2.31.0

